altos/scheme: Use 64-bit ints to track memory allocation stats
These are only collected for debug purposes, but can get quite large
if the interpreter runs for a while.

+++ src/scheme/ao_scheme.h
@@ -522,9 +522,9 @@ ao_scheme_poly_vector(ao_poly poly)
 
 /* memory functions */
 
-extern int ao_scheme_collects[2];
-extern int ao_scheme_freed[2];
-extern int ao_scheme_loops[2];
+extern uint64_t ao_scheme_collects[2];
+extern uint64_t ao_scheme_freed[2];
+extern uint64_t ao_scheme_loops[2];
 
 /* returns 1 if the object was already marked */
 int
